Reorganize repo; mega menu parent item caret; icon card section fixup
