+++ src/components/composite/oodQuoteSlider.js
