+++ src/components/simple/embedScript.js
