+++ src/scss/components/navigation/_mega-menu.scss
@@ -13,6 +13,22 @@
     line-height: 1.4;
     background-color: transparent;
     border: 0;
+
+    &::after {
+      @include grid-media('lg') {
+        @include margin(null null 1px 6px);
+        display: inline-block;
+        content: " ";
+        background: url("#{$su-image-path}/caret-down-black.svg") no-repeat 0 0;
+        background-size: 100%;
+        position: relative;
+        right: 0;
+        top: 0;
+        height: 11px; // Small svg gets chopped off if use em instead of px
+        width: 11px;
+        transition: transform 0.3s ease-out;
+      }
+    }
   }
 
   &__link {
